- Don't bother with move semantics on default constructed temporaries in TArray. - Cleared VS2005 warnings.

+++ src/id_sd.cpp
@@ -1331,7 +1331,7 @@ SD_PauseMusic(void)
 	if (music != NULL && Mix_PlayingMusic() == 1)
 	{
 		Mix_PauseMusic();
-		return Mix_GetMusicPCMPosition();
+		return (int)Mix_GetMusicPCMPosition();
 	}
 	return 0;
 }
